+++ src/common/Browser.ts
@@ -417,7 +417,8 @@ export class Browser extends EventEmitter {
   }
 
   /**
-   * Creates a {@link Page} in the default browser context.
+   * Promise which resolves to a new {@link Page} object. The Page is created in
+   * a default browser context.
    */
   async newPage(): Promise<Page> {
     return this._defaultContext.newPage();
@@ -432,7 +433,7 @@ export class Browser extends EventEmitter {
       url: 'about:blank',
       browserContextId: contextId || undefined,
     });
-    const target = await this._targets.get(targetId);
+    const target = this._targets.get(targetId);
     assert(
       await target._initializedPromise,
       'Failed to create target for page'
@@ -479,7 +480,7 @@ export class Browser extends EventEmitter {
     const { timeout = 30000 } = options;
     const existingTarget = this.targets().find(predicate);
     if (existingTarget) return existingTarget;
-    let resolve;
+    let resolve: (value: Target | PromiseLike<Target>) => void;
     const targetPromise = new Promise<Target>((x) => (resolve = x));
     this.on(BrowserEmittedEvents.TargetCreated, check);
     this.on(BrowserEmittedEvents.TargetChanged, check);
@@ -722,9 +723,8 @@ export class BrowserContext extends EventEmitter {
     permissions: Permission[]
   ): Promise<void> {
     const protocolPermissions = permissions.map((permission) => {
-      const protocolPermission = WEB_PERMISSION_TO_PROTOCOL_PERMISSION.get(
-        permission
-      );
+      const protocolPermission =
+        WEB_PERMISSION_TO_PROTOCOL_PERMISSION.get(permission);
       if (!protocolPermission)
         throw new Error('Unknown permission: ' + permission);
       return protocolPermission;