+++ src/common/ExecutionContext.ts
@@ -267,10 +267,8 @@ export class ExecutionContext {
         error.message += ' Are you passing a nested JSHandle?';
       throw error;
     }
-    const {
-      exceptionDetails,
-      result: remoteObject,
-    } = await callFunctionOnPromise.catch(rewriteError);
+    const { exceptionDetails, result: remoteObject } =
+      await callFunctionOnPromise.catch(rewriteError);
     if (exceptionDetails)
       throw new Error(
         'Evaluation failed: ' + helper.getExceptionMessage(exceptionDetails)