+++ src/common/WebWorker.ts
@@ -95,7 +95,7 @@ export class WebWorker extends EventEmitter {
       this._executionContextCallback(executionContext);
     });
 
-    // This might fail if the target is closed before we recieve all execution contexts.
+    // This might fail if the target is closed before we receive all execution contexts.
     this._client.send('Runtime.enable').catch(debugError);
     this._client.on('Runtime.consoleAPICalled', (event) =>
       consoleAPICalled(
@@ -151,7 +151,7 @@ export class WebWorker extends EventEmitter {
   /**
    * The only difference between `worker.evaluate` and `worker.evaluateHandle`
    * is that `worker.evaluateHandle` returns in-page object (JSHandle). If the
-   * function passed to the `worker.evaluateHandle` returns a [Promise], then
+   * function passed to the `worker.evaluateHandle` returns a `Promise`, then
    * `worker.evaluateHandle` would wait for the promise to resolve and return
    * its value. Shortcut for
    * `await worker.executionContext()).evaluateHandle(pageFunction, ...args)`